+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 * **Breakage**: Dropping support for python 3.2 due to lack of botocore support
 * **Breakage**: Changing the ``list``, ``dict``, and ``bool`` data types to use native DynamoDB types instead of JSON serializing.
 * Feature: Can now do 'contains' filters on lists
+* Feature: Fields support multiple validation checks
 
 Data type changes are due to an `update in the DynamoDB API
 <https://aws.amazon.com/blogs/aws/dynamodb-update-json-and-more/>`_