+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changelog
 0.4.1
 -----
 * **Warning**: Stored datetime objects will now be timezone-aware (:sha:`a7c253d`)
+* **Warning**: Stored datetime objects will now keep their microseconds (:sha:`fffe92c`)
 
 0.4.0
 -----